+++ src/init.c
@@ -92,7 +92,7 @@ extern SEXP vctrs_validate_name_repair_arg(SEXP);
 extern SEXP vctrs_validate_minimal_names(SEXP, SEXP);
 extern r_obj* ffi_vec_as_names(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_is_partial(r_obj*);
-extern SEXP vctrs_is_list(SEXP);
+extern r_obj* ffi_obj_is_list(r_obj*);
 extern SEXP vctrs_try_catch_callback(SEXP, SEXP);
 extern r_obj* ffi_is_coercible(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_as_subscript(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*);
@@ -275,7 +275,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_validate_minimal_names",          (DL_FUNC) &vctrs_validate_minimal_names, 2},
   {"ffi_vec_as_names",                      (DL_FUNC) &ffi_vec_as_names, 4},
   {"ffi_is_partial",                        (DL_FUNC) &ffi_is_partial, 1},
-  {"vctrs_is_list",                         (DL_FUNC) &vctrs_is_list, 1},
+  {"ffi_obj_is_list",                       (DL_FUNC) &ffi_obj_is_list, 1},
   {"vctrs_try_catch_callback",              (DL_FUNC) &vctrs_try_catch_callback, 2},
   {"ffi_is_coercible",                      (DL_FUNC) &ffi_is_coercible, 4},
   {"ffi_as_subscript",                      (DL_FUNC) &ffi_as_subscript, 5},